requirements.txt
numpy
pandas
torch
transformers
scikit-learn
matplotlib
anthropic
scipy
python-dotenv
openai